@@ -19,31 +22,27 @@
             "email": "fernando.isidro@ocelotlstudio.com"
         }
     ],
+    "homepage": "https://github.com/phpcfdi/csf-scraper",
     "support": {
-        "source": "https://github.com/phpcfdi/csf-scraper",
-        "issues": "https://github.com/phpcfdi/csf-scraper/issues"
-    },
-    "prefer-stable": true,
-    "config": {
-        "optimize-autoloader": true,
-        "preferred-install": {
-            "*": "dist"
-        }
+        "issues": "https://github.com/phpcfdi/csf-scraper/issues",
+        "source": "https://github.com/phpcfdi/csf-scraper"
     },
     "require": {
-        "php": ">=8.0",
-        "ext-mbstring": "*",
-        "ext-json": "*",
+        "php": ">=8.2",
         "ext-curl": "*",
+        "ext-json": "*",
+        "ext-mbstring": "*",
         "guzzlehttp/guzzle": "^7.4",
         "phpcfdi/rfc": "^1.1",
-        "symfony/css-selector": ">=6.0",
-        "symfony/dom-crawler": ">=6.0",
-        "symfony/process": ">=6.0"
+        "psr/http-client": "^1.0",
+        "symfony/css-selector": "^7.0",
+        "symfony/dom-crawler": "^7.0",
+        "symfony/process": "^7.0"
     },
     "require-dev": {
-      "phpunit/phpunit": "^9.5"
+        "phpunit/phpunit": "^11.5"
     },
+    "prefer-stable": true,
     "autoload": {
         "psr-4": {
             "PhpCfdi\\CsfScraper\\": "src/"
