@@ -6,7 +6,6 @@ namespace PhpCfdi\CsfScraper;
 
 use GuzzleHttp\Client;
 use GuzzleHttp\ClientInterface;
-use GuzzleHttp\Exception\GuzzleException;
 use PhpCfdi\CsfScraper\Exceptions\PdfReader\CifFromPdfNotFoundException;
 use PhpCfdi\CsfScraper\Exceptions\PdfReader\RfcFromPdfNotFoundException;
 use PhpCfdi\CsfScraper\Interfaces\ScraperInterface;
@@ -14,6 +13,7 @@ use PhpCfdi\CsfScraper\PdfReader\CsfExtractor;
 use PhpCfdi\CsfScraper\PdfReader\PdfToText;
 use PhpCfdi\Rfc\Exceptions\InvalidExpressionToParseException;
 use PhpCfdi\Rfc\Rfc;
+use Psr\Http\Client\ClientExceptionInterface;
 
 /**
  * Main class to obtain the data from a "Persona Moral" or a "Persona Física" from SAT website
@@ -22,7 +22,7 @@ class Scraper implements ScraperInterface
 {
     public static string $url = 'https://siat.sat.gob.mx/app/qr/faces/pages/mobile/validadorqr.jsf?D1=10&D2=1&D3=%s_%s';
 
-    public function __construct(private ClientInterface $client)
+    public function __construct(private readonly ClientInterface $client)
     {
     }
 
@@ -49,7 +49,7 @@ class Scraper implements ScraperInterface
      */
     public function obtainFromRfcAndCif(Rfc $rfc, string $idCIF): PersonaMoral|PersonaFisica
     {
-        $uri = sprintf(self::$url, $idCIF, $rfc->getRfc());
+        $uri = sprintf(static::$url, $idCIF, $rfc->getRfc());
         $html = $this->obtainHtml($uri);
         $person = (new DataExtractor($html))->extract($rfc->isFisica());
         $person->setRfc($rfc->getRfc());
@@ -61,12 +61,11 @@ class Scraper implements ScraperInterface
      * Obtain a PersonaMoral or PersonaFisica object with the information from SAT website
      * The RFC and CIF is taken from the "Constancia de Situación Fiscal" PDF file
      *
-     * @param string $path
-     * @return PersonaMoral|PersonaFisica
+     * @throws Exceptions\CifDownloadException
      * @throws Exceptions\CifNotFoundException
-     * @throws Exceptions\PdfReader\CifFromPdfNotFoundException
+     * @throws CifFromPdfNotFoundException
      * @throws Exceptions\PdfReader\EmptyPdfContentException
-     * @throws Exceptions\PdfReader\RfcFromPdfNotFoundException
+     * @throws RfcFromPdfNotFoundException
      * @throws InvalidExpressionToParseException
      */
     public function obtainFromPdfPath(string $path): PersonaMoral|PersonaFisica
@@ -88,8 +87,6 @@ class Scraper implements ScraperInterface
     /**
      * Helper method to extract the text information from PDF to plain text
      *
-     * @param string $path
-     * @return string
      * @throws Exceptions\PdfReader\PdfToTextConvertException when call to pdftotext fail
      */
     protected function pdfToTextContent(string $path): string
@@ -106,10 +103,10 @@ class Scraper implements ScraperInterface
     protected function obtainHtml(string $uri): string
     {
         try {
-            $request = $this->client->request('GET', $uri);
-        } catch (GuzzleException $exception) {
+            $response = $this->client->request('GET', $uri);
+        } catch (ClientExceptionInterface $exception) {
             throw new Exceptions\CifDownloadException($uri, $exception);
         }
-        return (string) $request->getBody();
+        return (string) $response->getBody();
     }
 }
