@@ -14,9 +14,6 @@ final class ScraperTest extends TestCase
 {
     public function test_create_with_specific_openssl_cipher_list(): void
     {
-        /**
-         * Problem: The scraper does not have a method to get the client, obtained by reflection
-         */
         $scraper = Scraper::create();
         $client = $scraper->getClient();
         $this->assertInstanceOf(Client::class, $client);
@@ -27,7 +24,7 @@ final class ScraperTest extends TestCase
          */
         $this->assertSame(
             [CURLOPT_SSL_CIPHER_LIST => 'DEFAULT@SECLEVEL=1'],
-            $client->getConfig('curl')
+            $client->getConfig('curl'),
         );
     }
 }
